Move Wetherfield-specific bits to own file
